requirements.txt
fastapi
uvicorn
httpx
pydantic
qdrant-client
python-dotenv
strands-agents
ollama
anthropic
